# services/skillpod-svc/docs/api/data-containment.openapi.yaml
openapi: 3.0.3
info:
  title: SkillPod Data Containment API
  description: |
    API for managing SkillPod VDI (Virtual Desktop Infrastructure) security, 
    including pod lifecycle management, session monitoring, data loss prevention (DLP), 
    and real-time policy enforcement.

    ## Security Features
    - Clipboard copy/paste blocking/logging
    - File download/upload controls
    - Screen capture prevention with watermarking
    - USB device access policies
    - Print policy enforcement
    - Real-time DLP scanning

    ## Authentication
    All endpoints require a valid JWT token with appropriate tenant and user claims.
  version: 1.0.0
  contact:
    name: Skillancer Platform Team
    email: platform@skillancer.com
  license:
    name: Proprietary
    url: https://skillancer.com/terms
servers:
  - url: /api/v1/skillpod
    description: SkillPod Service API
...
paths:
  /pods:
    post:
      tags:
        - Pods
      summary: Create a new pod
      description: |
        Creates a new SkillPod VDI pod with the specified configuration.
        A Kasm workspace is provisioned automatically.
      operationId: createPod
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePodRequest'
      responses:
        '201':
          description: Pod created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PodResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags:
        - Pods
      summary: List all pods
      description: Returns a paginated list of pods for the current tenant
      operationId: listPods
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          description: Filter by pod status
          schema:
            type: string
            enum: [PROVISIONING, RUNNING, PAUSED, STOPPING, TERMINATED, FAILED]
        - name: userId
          in: query
          description: Filter by user ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of pods
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PodListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /pods/{podId}:
    get:
      tags:
        - Pods
      summary: Get pod details
      description: Returns detailed information about a specific pod including Kasm status
      operationId: getPod
      parameters:
        - $ref: '#/components/parameters/PodId'
      responses:
        '200':
          description: Pod details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PodDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /pods/{podId}/policy:
    put:
      tags:
        - Pods
      summary: Apply security policy to pod
      description: Applies a security policy to the pod and updates Kasm workspace settings
      operationId: applyPodPolicy
      parameters:
        - $ref: '#/components/parameters/PodId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - policyId
              properties:
                policyId:
                  type: string
                  format: uuid
                  description: The security policy ID to apply
      responses:
        '200':
          description: Policy applied successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyApplyResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /pods/{podId}/action:
    post:
      tags:
        - Pods
      summary: Perform action on pod
      description: Performs lifecycle actions on the pod (pause, resume, terminate, restart)
      operationId: podAction
      parameters:
        - $ref: '#/components/parameters/PodId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum: [pause, resume, terminate, restart]
                reason:
                  type: string
                  description: Reason for the action (required for terminate)
      responses:
        '200':
          description: Action performed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /pods/{podId}/watermark:
    get:
      tags:
        - Pods
      summary: Get watermark configuration
      description: Returns the watermark configuration for the pod
      operationId: getPodWatermark
      parameters:
        - $ref: '#/components/parameters/PodId'
      responses:
        '200':
          description: Watermark configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatermarkResponse'
...
components:
  parameters:
    PodId:
      name: podId
      in: path
      required: true
      description: The pod ID
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
    CreatePodRequest:
      type: object
      required:
        - imageId
      properties:
        imageId:
          type: string
          description: The Kasm image ID to use for the workspace
        policyId:
          type: string
          format: uuid
          description: Security policy to apply (uses default if not specified)
        name:
          type: string
          description: Optional name for the pod
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata for the pod
    PodResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PROVISIONING, RUNNING, PAUSED, STOPPING, TERMINATED, FAILED]
        kasmId:
          type: string
        connectionUrl:
          type: string
          format: uri
        policyId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    PodListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PodResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PodDetailResponse:
      allOf:
        - $ref: '#/components/schemas/PodResponse'
        - type: object
          properties:
            kasmStatus:
              type: object
              properties:
                status:
                  type: string
                uptime:
                  type: integer
                container_id:
                  type: string
            policy:
              $ref: '#/components/schemas/PolicySummary'
    PolicyApplyResponse:
      type: object
      properties:
        success:
          type: boolean
        policyId:
          type: string
          format: uuid
        appliedAt:
          type: string
          format: date-time
    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
        action:
          type: string
        newStatus:
          type: string
        timestamp:
          type: string
          format: date-time
...
    WatermarkResponse:
      type: object
      properties:
        enabled:
          type: boolean
        text:
          type: string
        opacity:
          type: number
        position:
          type: string
          enum: [center, corner, tiled]
        includedInfo:
          type: array
          items:
            type: string
            enum: [userId, sessionId, timestamp, ipAddress]
...
    PolicySummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        clipboardPolicy:
          type: string
        fileDownloadAllowed:
          type: boolean
        fileUploadAllowed:
          type: boolean
        screenCaptureBlocked:
          type: boolean
        printingPolicy:
          type: string
        usbPolicy:
          type: string
        dlpEnabled:
          type: boolean
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []